All programs now have a proper entry class that is derived of off Tj3AppBase
