+++ spec/Tj3Daemon_spec.rb
@@ -12,64 +12,20 @@
 #
 
 require 'rubygems'
-require 'support/StdIoWrapper'
-require 'taskjuggler/apps/Tj3Daemon'
-require 'taskjuggler/apps/Tj3Client'
+require 'support/DaemonControl'
 
 class TaskJuggler
 
   describe Tj3Daemon do
 
-    include StdIoWrapper
+    include DaemonControl
 
     before(:each) do
-      (f = File.new('taskjuggler.rc', 'w')).write(<<'EOT'
-_global:
-  authKey: 'secret_key'
-  port: 0
-  _log:
-    outputLevel: 3
-    logLevel: 3
-EOT
-                                                 )
-      f.close
-
-      if (pid = fork).nil?
-        at_exit { exit! }
-        $stdout.reopen('stdout.log', 'w')
-        $stderr.reopen('stderr.log', 'w')
-        res = stdIoWrapper do
-          Tj3Daemon.new.main(%w( --silent ))
-        end
-        raise "Failed to start tj3d: #{res.stdErr}" if res.returnValue != 0
-        exit!
-      else
-        # Wait for the daemon to get online.
-        i = 0
-        while !File.exists?('.tj3d.uri') && i < 10
-          sleep 0.5
-          i += 1
-        end
-        raise 'Daemon did not start properly' if i == 10
-      end
-      0
+      startDaemon
     end
 
     after(:each) do
-      res = stdIoWrapper do
-        Tj3Client.new.main(%w( --silent --unsafe terminate ))
-      end
-      raise "tj3d termination failed: #{res.stdErr}" if res.returnValue != 0
-      i = 0
-      while File.exists?('.tj3d.uri') && i < 10
-        sleep 0.5
-        i += 1
-      end
-      raise "Daemon did not terminate properly" if i == 10
-      # Cleanup file system again.
-      %w( taskjuggler.rc stdout.log stderr.log ).each do |file|
-        File.delete(file)
-      end
+      stopDaemon
     end
 
     it 'should be startable and stopable' do
